CI attempt 2

+++ clean_build_all.sh
@@ -1,5 +1,5 @@
 cleanProject () {
-	rm build/libs/*
+	[ ! -d build/libs ] || rm build/libs/*
 }
 
 ( cd projects/1.7 && cleanProject && gradlew build )
